fix: Solve query result parsing bug
When parsing the raw query results, the input value column often came in
the wrong order, one of the frustrating things string sorting can do.
Now the query dict uses integers for keys and the result keys are sorted
with keys=int,this should fix the problems. Also expanding cache size,
because I don't think there's a good enough reason not to do so.