feat: Makes reconciliation endpoint configurable
Making the reconciliation endpoint a configurable argument was a
valid request that I liked a lot, so there is now a
reconciliation_endpoint argument in the main reconcile() function,
that only defaults to Wikidata if no value is given.

This commit also changes the qid_type argument to be type_id, a more
generic name that should make it more intuitive to integrate to other
knowledge bases.

Addresses #4