feat: Add option to reconcile against triples
Added a new parameter to the reconcile function that allows the user to
reconcile against a specific triple, so if desired, the user can define
the has_property tuple with a property and its respective value to
reconcile against. This is extremely useful in biological scenarios,
which will be my main use. I could, for example, reconcile against
proteins of a specific taxon.

* Addresses issue #2